Add files via upload
With a huge product catalog that includes over 10,000 items, Global Superstore, an international online retailer, hopes to serve as a one-stop shop for its clients. The organization serves customers in 147 countries and works in multiple global markets. This project aimed to examine Global Superstore's 2014 sales data to spot patterns, opportunities, and difficulties.